UI improvements, show scores as percentages

+++ app/src/main/java/ocrtest/camera/heuristics/HeuristicOutput.kt
@@ -5,4 +5,4 @@ import com.google.common.collect.ImmutableMap
 /**
  * Output from a single Heuristic
  */
-data class HeuristicOutput(val scores: ImmutableMap<String, Int>)
+data class HeuristicOutput(val scores: ImmutableMap<String, Double>)
