+++ app/src/main/java/ocrtest/camera/heuristics/question_answer_search/GoogleSearchHelper.kt
@@ -13,23 +13,23 @@ class GoogleSearchHelper(val searchService : GoogleSearchService) {
     // About 9,870,000,000 results
     private val PATTERN = Pattern.compile("About ([0-9,]+) results")
 
-    fun getResultCount(query : String) : Observable<Int> {
+    fun getResultCount(query : String) : Observable<Double> {
         return searchService.search(query)
                 .subscribeOn(Schedulers.computation())
                 .map({responseBody -> extractMatchCount(responseBody.string())})
     }
 
-    private fun extractMatchCount(page: String) : Int {
+    private fun extractMatchCount(page: String) : Double {
         val matcher = PATTERN.matcher(page)
         if (matcher.find()) {
             try {
                 return Integer.parseInt(
                         matcher.group(0)
-                                .filter { char -> (char >= '0' && char <= '9') })
+                                .filter { char -> (char >= '0' && char <= '9') }).toDouble()
             } catch (exception : NumberFormatException) {
-                return Integer.MAX_VALUE
+                return Integer.MAX_VALUE.toDouble()
             }
         }
-        return 0
+        return 0.0
     }
 }
